# api/openapi.yaml
openapi: 3.1.0
info:
  title: blabby HTTP API
  version: 0.1.0
  description: |
    The HTTP API for blabby, a real-time chat service.

    Clients use HTTP for commands and queries (login, room discovery, membership,
    and sending messages) and a WebSocket for the live event stream. The WebSocket
    message contract is documented separately in `asyncapi.yaml`.

    ## Authentication

    Every endpoint requires an `Authorization: Bearer <jwt>` header **except**:

    - `POST /users`, `POST /users/verifications`, and
      `POST /users/verifications/resend` — the pre-login registration flow,
    - `POST /login`, which mints the token and is therefore public, and
    - `GET /ws`, where authentication is the first WebSocket frame after the
      upgrade rather than an HTTP header.

    A missing or malformed bearer header is rejected with `401 AUTH_MISSING_TOKEN`,
    an expired token with `401 AUTH_EXPIRED_TOKEN`, and any other invalid token
    with `401 AUTH_INVALID_TOKEN`.

    ## Error envelope

    Every application-generated JSON error response shares one envelope: a top-level
    `error` object holding a numeric `code`, a canonical `status` string, and a
    human-readable, client-safe `message`. A failed WebSocket upgrade is emitted by
    the WebSocket library as plain text instead. The `status` enum on `ErrorDetail`
    lists the full error taxonomy; each operation's `responses` section is the
    authoritative list of what that operation can actually return.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags:
        - Rooms
      summary: List the available rooms.
      description: |
        Returns one page of the catalogue of rooms a client can join, in
        creation order. `q` narrows the page to rooms whose display name
        contains the fragment (case-insensitive, literal — no wildcard
        syntax). Pages continue by echoing the response's `next` cursor back
        as `after`; a `null` `next` means the listing is exhausted.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            x-max-utf8-bytes: 64
          description: |
            A display-name search fragment, following the same character rules
            as room names: 1-64 bytes of printable characters. A blank value
            is treated as absent.
        - name: after
          in: query
          required: false
          schema:
            type: string
          description: |
            The keyset cursor: the `R…` id of the last room on the previous
            page, as returned in `next`. Only rooms created after it are
            listed. A cursor that matches no active room is rejected with a
            400; restart from the first page.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: The page size.
      responses:
        '200':
          description: One page of the room catalogue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomCataloguePage'
              example:
                rooms:
                  - id: RG000000004
                    name: General
                  - id: RH000000005
                    name: Random
                next: null
        '400':
          description: |
            `q` was not a valid name fragment, `after` was malformed or matched
            no active room, or `limit` was not an integer between 1 and 200.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4001
                  status: INVALID_REQUEST
                  message: after references an unknown room
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      operationId: createRoom
      tags:
        - Rooms
      summary: Create a room.
      description: |
        Creates an active room owned by the authenticated caller: the room, the
        caller's owner membership, and the founding `member_joined` timeline
        event commit together, so the room's timeline explains its owner like
        any later member. The new room appears in the caller's
        `GET /rooms/joined` immediately. Room names are display labels — not
        unique, any printable characters — and the returned `R…` code is the
        room's identity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
            example:
              name: Team Standup
      responses:
        '201':
          description: The room exists; the body carries its descriptor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDescriptor'
              example:
                id: RK000000042
                name: Team Standup
        '400':
          description: |
            The body was malformed, or `name` was missing, blank, longer than
            64 bytes, or contained control or invisible formatting characters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4001
                  status: INVALID_REQUEST
                  message: name must be 1-64 bytes of printable characters
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: The request body exceeded the 1 KiB limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4003
                  status: PAYLOAD_TOO_LARGE
                  message: request body exceeds maximum size
        '500':
          $ref: '#/components/responses/InternalError'
...
  /rooms/{id}/events:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      operationId: listRoomEvents
      tags:
        - Rooms
      summary: Read a page of the room's timeline.
      description: |
        Returns one newest-first page of the room's durable timeline — chat
        messages and `member_joined`/`member_left` system events interleaved
        in id order. The caller must be a current member. `q` narrows the page
        to messages whose text contains every whitespace-separated term of the
        fragment (full-text, CJK-aware; terms match literally — no query
        syntax). Older pages continue by echoing the response's `next` cursor
        back as `before`; a `null` `next` means the history is exhausted.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            x-max-utf8-bytes: 256
          description: |
            A message-text search fragment: 1-256 bytes of valid UTF-8. Every
            whitespace-separated term must appear in the message. When set,
            only `message` events are returned. A blank value is treated as
            absent.
        - name: before
          in: query
          required: false
          schema:
            type: string
          description: |
            The keyset cursor: the event id of the last entry on the previous
            page, as returned in `next`. Only older events are listed. A blank
            value is treated as absent.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: The page size.
      responses:
        '200':
          description: One page of the room timeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomEventsPage'
              example:
                events:
                  - id: '112233445566778899'
                    type: message
                    sender:
                      id: UA000000001
                      name: alice
                    text: hello, world
                    timestamp: 1718280000000
                  - id: '112233445566778898'
                    type: member_joined
                    user:
                      id: UA000000001
                      name: alice
                    timestamp: 1718279000000
                next: null
        '400':
          description: |
            The room identifier was invalid, `q` was not a valid search
            fragment, `before` was not a valid event id, or `limit` was not an
            integer between 1 and 200.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 4001
                  status: INVALID_REQUEST
                  message: before is not a valid event id
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: The caller is not a member of the room.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: 2001
                  status: ROOM_NOT_MEMBER
                  message: not a member of this room
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "A JWT issued by `POST /login`, sent as `Authorization: Bearer <jwt>`."
  parameters:
    RoomId:
      name: id
      in: path
      required: true
      description: |
        The opaque public room code: `R` followed by 10 Crockford Base32 symbols
        (e.g. `RG000000004`). Input is parsed leniently (case-insensitive, `O`→`0`,
        `I`/`L`→`1`). A malformed code is rejected with `400 INVALID_REQUEST`; a
        well-formed code that matches no active room is `404 ROOM_NOT_FOUND`. The
        internal numeric room id is never accepted here.
      schema:
        type: string
      example: RG000000004
...
  schemas:
...
    RoomCataloguePage:
      type: object
      required:
        - rooms
        - next
      properties:
        rooms:
          type: array
          description: One page of the catalogue, as descriptors in creation order.
          items:
            $ref: '#/components/schemas/RoomDescriptor'
        next:
          type:
            - string
            - 'null'
          description: |
            The `after` value for the next page — the id of the last room in
            `rooms` — or `null` when the listing is exhausted.
    RoomDescriptor:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: |
            The opaque public room code (`R…`) used in path parameters and event
            frames. The internal numeric room id is never exposed.
        name:
          type: string
          description: The room's human-readable display name.
    RoomEventsPage:
      type: object
      required:
        - events
        - next
      properties:
        events:
          type: array
          description: One newest-first page of the room timeline.
          items:
            $ref: '#/components/schemas/RoomEvent'
        next:
          type:
            - string
            - 'null'
          description: |
            The `before` value for the next (older) page — the id of the last
            event in `events` — or `null` when the history is exhausted.
    RoomEvent:
      type: object
      required:
        - id
        - type
        - timestamp
      properties:
        id:
          type: string
          description: |
            The event's id as a decimal string — the timeline cursor. Ids are
            programmatic identifiers, never human-shared.
        type:
          type: string
          enum:
            - message
            - member_joined
            - member_left
          description: |
            The event kind. A `message` carries `sender` and `text`;
            `member_joined`/`member_left` carry `user` — mirroring the
            WebSocket frames' vocabulary for people.
        sender:
          $ref: '#/components/schemas/EventPerson'
          description: The message's author. Present on `message` events only.
        user:
          $ref: '#/components/schemas/EventPerson'
          description: |
            The member who joined or left. Present on `member_joined` and
            `member_left` events only.
        text:
          type: string
          description: The message body. Present on `message` events only.
        timestamp:
          type: integer
          format: int64
          description: The event's server-assigned time in unix milliseconds.
    EventPerson:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: |
            The person's opaque public user code (`U…`). The internal numeric
            user id is never exposed.
        name:
          type: string
          description: The person's current display name.
...
    CreateRoomRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          x-max-utf8-bytes: 64
          description: |
            The room's display name: trimmed, non-blank, at most 64 bytes of
            printable characters (spaces included; control and invisible
            formatting characters are rejected). It is a label, not an
            address, and names are deliberately non-unique.
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - status
        - message
      properties:
        code:
          type: integer
          description: The numeric error code paired with `status`.
        status:
          type: string
          description: |
            The canonical status string. This enum is the complete error
            taxonomy, not the subset any single endpoint returns; consult each
            operation's `responses` for the codes it can actually produce.
          enum:
            - AUTH_INVALID_TOKEN
            - AUTH_EXPIRED_TOKEN
            - AUTH_MISSING_TOKEN
            - AUTH_ACCOUNT_PENDING
            - ROOM_NOT_MEMBER
            - ROOM_ALREADY_MEMBER
            - ROOM_NOT_FOUND
            - ROOM_OWNER_CANNOT_LEAVE
            - ROOM_PERMISSION_DENIED
            - RATE_LIMIT_EXCEEDED
            - VERIFICATION_RATE_LIMITED
            - INVALID_REQUEST
            - MISSING_FIELD
            - PAYLOAD_TOO_LARGE
            - INVALID_EMAIL
            - WEAK_PASSWORD
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - EMAIL_ALREADY_REGISTERED
            - HANDLE_ALREADY_TAKEN
            - VERIFICATION_INVALID
        message:
          type: string
          description: A human-readable, client-safe explanation. Never leaks internal detail.
  responses:
...
    Unauthorized:
      description: |
        The bearer token is missing, malformed, expired, or otherwise invalid
        (`AUTH_MISSING_TOKEN`, `AUTH_EXPIRED_TOKEN`, or `AUTH_INVALID_TOKEN`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 1003
              status: AUTH_MISSING_TOKEN
              message: missing authorization header
...
    RoomNotFound:
      description: |
        The room code is well-formed but matches no active room (the room does not
        exist or is no longer active).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 2003
              status: ROOM_NOT_FOUND
              message: room not found
    ServiceUnavailable:
      description: A backend dependency was temporarily unreachable. The request may be retried.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 5002
              status: SERVICE_UNAVAILABLE
              message: failed to reach user grain
    InternalError:
      description: An unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 5001
              status: INTERNAL_ERROR
              message: authentication unavailable